# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PCast REST-API",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/feeds": {
            "get": {
                "description": "Retrieve all feeds from the store",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Get all feeds",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Feed"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new feed with the data provided in the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Create a new feed",
                "parameters": [
                    {
                        "description": "Feed data",
                        "name": "feed",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateFeedRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Feed"
                        }
                    }
                }
            }
        },
        "/feeds/{id}": {
            "delete": {
                "description": "Delete a feed with the given ID",
                "tags": [
                    "feeds"
                ],
                "summary": "Delete a feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed deleted successfully"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateFeedRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "model.Feed": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        }
    }
}